+++ example/bin/numerical_iterable_example.dart
@@ -1,11 +1,21 @@
 import 'package:list_operators/list_operators.dart';
 
 void main() {
-  final b = [11, 12, 13];
+  final a = [10.0, 11.0, 12.0];
+  final b = <num>[11.0, 12, 13];
+  final c = [0, 1, 2];
 
   // Operators and method that work with Iterable<num>
+  print('Minimum: a.min()');
+  print(a.min());
+  print('');
+
   print('Minimum: b.min()');
   print(b.min());
+  print('');
+
+  print('Minimum: c.min()');
+  print(c.min());
   print('');
 
   print('Maximum: b.max()');
@@ -24,6 +34,16 @@ void main() {
   print(b.stdDev());
   print('');
 
+  print('Sum: a.sum()');
+  print(a.sum());
+  print('');
+
   print('Sum: b.sum()');
   print(b.sum());
+  print('');
+
+  print('Sum: c.sum()');
+  print(c.sum());
+
+  print(c.toListOfDouble());
 }