+++ lib/src/extensions/double_operators.dart
@@ -15,11 +15,6 @@ extension DoubleOperators on List<double> {
     return List<double>.generate(length, (i) => this[i] * scalar);
   }
 
-  /// Returns a new list containing the elements of `this` converted to `int`.
-  List<int> toInt() {
-    return List<int>.generate(length, (i) => this[i].toInt());
-  }
-
   /// Returns `true` if the equality
   /// `this(i) == other(i)` holds for each index `i`.
   bool match(List<double> other, {double precision = 1e-12}) {
@@ -40,9 +35,9 @@ extension DoubleIterableOperators on Iterable<double> {
   /// Returns the sum of the entries.
   ///
   /// The iterable must not be empty.
-  num sum() {
+  double sum() {
     mustHaveElements();
-    return fold<num>(0.0, (sum, current) => sum + current);
+    return fold<double>(0.0, (sum, current) => sum + current);
   }
 
   /// Returns the minimum value.